Fix PR #2 CI: extract sponsor validation, queue sync, and budgets.
Move sponsor schema guards to shared/sponsors_validate.py to bring server.py
under the line cap; clear stale isQueueFull when client chips end; skip
gitignored admin in ad-location sync test; fix ruff; refresh bundle ceilings.
